web service call

+++ dqs/models/systemcfg.go
@@ -46,6 +46,6 @@ type MailConfig struct {
 //地震报送控制数据
 type QuakeReportConfig struct {
 	ServiceURL string
-	SleepTime  int
-	TimeSpan   int
+	SleepTime  int64
+	TimeSpan   int64
 }
